backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
sqlalchemy[asyncio]>=2.0,<3.0
asyncpg>=0.30,<1.0
pyjwt>=2.9,<3.0
httpx>=0.28,<1.0
sse-starlette>=2.1,<3.0
arq>=0.26,<1.0
boto3>=1.35,<2.0
alembic>=1.14,<2.0
pydantic-settings>=2.7,<3.0
redis>=5.2,<6.0
structlog>=24.4,<26.0
python-multipart>=0.0.18,<1.0
